+++ codebase/apps/titan/src/EchoTops/EchoTops.cc
@@ -276,7 +276,7 @@ int EchoTops::_processFile(DsMdvx &inMdvx)
   
   // set output compression
 
-  for (int i = 0; i < outMdvx.getNFields(); i++) {
+  for (size_t i = 0; i < outMdvx.getNFields(); i++) {
     MdvxField *field = outMdvx.getFieldByNum(i);
     field->requestCompression(Mdvx::COMPRESSION_GZIP);
   }
@@ -284,7 +284,7 @@ int EchoTops::_processFile(DsMdvx &inMdvx)
   // add any chunks
   
   outMdvx.clearChunks();
-  for (int i = 0; i < inMdvx.getNChunks(); i++) {
+  for (size_t i = 0; i < inMdvx.getNChunks(); i++) {
     MdvxChunk *chunk = new MdvxChunk(*inMdvx.getChunkByNum(i));
     outMdvx.addChunk(chunk);
   }
